% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{act_crime_stat_2024_09_30}
\alias{act_crime_stat_2024_09_30}
\title{Crime Statistics for Australian Capital Territory, Australia from 1 Jan 2014 to 30 Sep 2024}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 101308 rows and 8 columns.
}
\usage{
act_crime_stat_2024_09_30
}
\description{
A tibble containing 101,308 rows and 8 columns including
`district`, `suburb`, `crime_code`, `crime_type`, `year`, `quarter`,
`month_range` and `count`.
}
\keyword{datasets}
