% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_act_crime_stat.R
\name{get_act_crime_stat}
\alias{get_act_crime_stat}
\title{Download Crime Statistics for Australian Capital Territory, Australia}
\usage{
get_act_crime_stat()
}
\value{
A tibble containing 8 columns including
`district`, `suburb`, `crime_code`, `crime_type`, `year`, `quarter`,
`month_range` and `count`.
}
\description{
Download Crime Statistics for Australian Capital Territory, Australia.
See https://www.data.act.gov.au/Justice-Safety-and-Emergency/ACT-Crime-Statistics/2egm-dieb/about_data
for more details.
}
